REDCapSync Encapsulated Datsaset Object

R6 project object for REDCapSync

Value

An R6ClassGenerator which is used internally to create or load a dataset object for the user

See also

Public fields

data

list of data where names are forms

metadata

list of metadata

records

data.frame of records with timestamps

users

data.frame of users with timestamps

log

data.frame of log

comments

data.frame of comments

Active bindings

project_name

Read-only character string of project_name as assigned with setup_project.

dataset_name

Read-only character string of dataset_name

n_records

Read-only integer of n_records

Methods

Public methods

REDCapSyncDataset$new()

REDCapSyncDataset$print()

Method new()

The end user will not see dataset$new(). This is handled internally. Users should construct objects using REDCapSyncProject.

Usage

REDCapSyncDataset$new(project, dataset_name)

Arguments

project

a list object meant to be stored internally within R6

dataset_name

dataset name setup with project$add_dataset

Method print()

Print some key dataset information

Usage

REDCapSyncDataset$print()