Deidentify the REDCap Database

Removes or masks identifying information from the REDCap database (project). This can be done either based on the identifier field in the metadata or by specifying custom identifiers.

Usage

deidentify_project(project, identifiers, drop_free_text = FALSE)

Arguments

project

A validated project object containing REDCap project data and settings. Generated using load_project or setup_project

identifiers

Optional character vector of column names that should be excluded from the project. If not provided, fields where project$metadata$fields$identifier == "y" will be used as the default.

drop_free_text

Logical. If TRUE, columns containing free text will also be excluded from the project. Default is FALSE.

Value

A project object with deidentified forms.

Details

This function modifies the project object to exclude specified identifiers or any columns flagged as identifiers in the metadata. Free-text fields can also be optionally removed, ensuring the resulting dataset complies with deidentification standards.

See also