Select REDCap Records from project

This function filters the records in the project object by specified criteria, such as field names, form names, and optional filtering based on a specific field and its values. It returns a modified project object containing only the records that match the filter criteria.

Usage

generate_project_summary_test(
  project,
  transform,
  filter_field,
  filter_choices,
  filter_list = NULL,
  filter_strict = TRUE,
  field_names = NULL,
  form_names = NULL,
  no_duplicate_cols = FALSE,
  deidentify = TRUE,
  drop_free_text = FALSE,
  clean = TRUE,
  drop_blanks = TRUE,
  include_metadata = TRUE,
  annotate_metadata = TRUE,
  include_record_summary = TRUE,
  include_users = TRUE,
  include_log = TRUE
)

Arguments

project

A validated project object containing REDCap project data and settings. Generated using load_project or setup_project

transform

Logical. Whether to transform the data in the subset. Default is TRUE.

filter_field

Character. The name of the field in the database to filter on.

filter_choices

Vector. The values of filter_field used to define the subset.

filter_list

Vector. The values of filter_field used to define the subset.

filter_strict

Logical. If TRUE, all forms will be filtered by criteria. If FALSE, will convert original filter to id column and filter all other forms by that record. Default is TRUE.

field_names

Character vector. Names of specific fields to include in the subset. Default is NULL, which includes all fields.

form_names

Character vector. Names of forms to include in the subset. Default is NULL, which includes all forms.

no_duplicate_cols

A logical flag (TRUE or FALSE). If TRUE, the function will avoid including duplicate columns in the output. Defaults to FALSE.

deidentify

Logical. Whether to deidentify the data in the subset. Default is TRUE.

drop_free_text

Logical. If TRUE, columns containing free text will also be excluded from the project. Default is FALSE.

clean

Logical. If TRUE, the data will be cleaned before summarizing. Default is TRUE.

drop_blanks

Logical. If TRUE, records with blank fields will be dropped. Default is TRUE.

include_metadata

Logical. If TRUE, metadata will be included in the summary. Default is TRUE.

annotate_metadata

Logical. If TRUE, metadata will be annotated in the summary. Default is TRUE.

include_record_summary

Logical. If TRUE, a record summary will be included in the generated summary. Default is TRUE.

include_users

Logical. If TRUE, user-related information will be included in the summary. Default is TRUE.

include_log

Logical. If TRUE, the log of changes will be included in the summary. Default is TRUE.

Value

A modified project object with filtered records and columns based on the provided criteria.

Details

This function filters the data in the project object according to the specified form and field names and optional filter criteria. If no field names or form names are provided, it defaults to using all fields and forms in the database. The function uses the helper filter_form_list to apply the filtering logic to the project$data list.